requirements.txt
celery==5.3.6
Django==5.0.4
django-simple-history==3.7.0
djongo5==1.3.9
docker==7.0.0
pre-commit==3.7.0
pydantic==2.6.4
pymongo==3.12.3
pytz==2024.1
redis==5.0.3
tldextract==5.1.2